refactor a bit to sub directories
