+++ main.py
@@ -1,34 +1,13 @@
 from datetime import datetime, timedelta
-import os.path
 import pytz
 import matplotlib.pyplot as plt
-from google.auth.transport.requests import Request
-from google.oauth2.credentials import Credentials
-from google_auth_oauthlib.flow import InstalledAppFlow
+from function.credentials import get_creds
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 import os
 
 os.environ['http_proxy'] = 'http://127.0.0.1:7890'
 os.environ['https_proxy'] = 'https://127.0.0.1:7890'
-
-SCOPES = ["https://www.googleapis.com/auth/calendar.readonly"]
-
-def get_creds():
-    """Retrieve user credentials for Google Calendar API."""
-    creds = None
-    if os.path.exists("token.json"):
-        creds = Credentials.from_authorized_user_file("token.json", SCOPES)
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file("credentials.json", SCOPES)
-            creds = flow.run_local_server(port=0)
-        with open("token.json", "w") as token:
-            token.write(creds.to_json())
-    return creds
-
 
 def get_calendar_list(service):
     """Get list of calendars from the Google Calendar API."""
